Correctly handle config fields in multi component packages
