+++ internal/cmd/tree.go
@@ -72,7 +72,13 @@ func (t *Tree) RenderPackage(ctx context.Context, srcPath string, opts ...Render
 		return "", fmt.Errorf("loading package contents from folder: %w", err)
 	}
 
-	pkg, err := packagecontent.PackageFromFiles(ctx, t.scheme, files, cfg.Component)
+	// TODO: show all components in the tree
+	filesByComponent, err := packagecontent.SplitFilesByComponent(ctx, t.scheme, files, cfg.Component)
+	if err != nil {
+		return "", fmt.Errorf("parsing package contents: %w", err)
+	}
+
+	pkg, err := packagecontent.PackageFromFiles(ctx, t.scheme, filesByComponent[cfg.Component])
 	if err != nil {
 		return "", fmt.Errorf("parsing package contents: %w", err)
 	}
